Include Driver and mysql in main project, remove copy of files
