--- bd2/README.md
@@ -1,2 +1,0 @@
-# bd2
-University project for subject Data Base 2
